requirements.txt
pyserial
pywin32